requirements.txt
Django==1.7.2
Django-Select2==4.2.2
django-bootstrap3==5.0.3
django-jsonview==0.4.3
django-sorting-bootstrap==1.0.2
django-su==0.4.7
yweather==0.1
psycopg2==2.5.4
djangorestframework==2.4.3
django-allauth==0.18.0
oauthlib==0.7.2
python-openid==2.2.5
python3-openid==3.0.5
requests==2.5.1
requests-oauthlib==0.4.2
django-filter==0.9.1
django-jquery==1.9.1
django-cors-headers==1.0.0
django-ember==0.3.1
django.js==0.8.1
pygeocoder==1.2.5